Updates with first analysis consoles for the october 24 measurements.
New voltage comparison, with important adding of a filer for the diode signal. This changed the results I obtained and compared earlier for the 128 line arrays.

Added mappings of the 2 new measured matrix arrays. (As before the channel numbering needs to be translated) Consoles for movie of moved logo and beam increase.

Sorting of the older consoles in to corresponding folders.

Increase of the standard plot fontsize to uniformly 12.