Some additions of logbooks missing so far (also of recent measurements from March 25).
Some adjustments of plot_standards.py to avoid problems in several graph formats.

Bigger change: I added the signal conversion from AMS units to current of the AMS circuit. (Details of the formula in a coming results presentation). I also adapted some plot codes to respect this.
The signal_conversion is now a function of the Analyzer class and can be called for each instance of the class. The necessary parameters clock, gain and for plotting reasons: scale are added as class variables. Standard scaling of current will be to pA.
Class internally I will keep the AMS values until the maps of the data are created. This way no conflict arises from merging scaled signals and potentially unscaled dark levels. Externally I changed functions in standard_processes, but some control here might be necessary.