Changelog:
New Consoles...
for different parameters connected with the wedge geometry to judge their impact on the exp/sim comparison.
for the simulated results of the energy extrapolation.

New Style Consoles added to design graphics for, e.g., publication in the future. So far the Style1_SelectedImages.py is filled and used.

Minor changes in plotting code, added (float, int) in helper functions data to axis and vice versa trafo functions.

New helper function in main to automatically get a string with the significant digits when giving a number and its uncertainty.
-> Will think about implementing analysis with uncertainties package in the future!